Removing the scripts folder and clearing out some more crap.
Neither of the scripts are at all useful to me and I don't see the point
in having the UnitTest++ directory, so now it's gone.

Also removed the "replace.py" script. Still not sure what it's for.
If I bothered to look in the Makefile, I'd probably get my answer. Not
going to look though, because I like myself as a person too much to look
at a Makefile on purpose.